+++ mobile/android/geckoview_example/src/main/java/org/mozilla/geckoview_example/GeckoViewActivity.java
@@ -8,6 +8,7 @@ package org.mozilla.geckoview_example;
 import org.json.JSONObject;
 
 import org.mozilla.geckoview.AllowOrDeny;
+import org.mozilla.geckoview.Autocomplete;
 import org.mozilla.geckoview.BasicSelectionActionDelegate;
 import org.mozilla.geckoview.ContentBlocking;
 import org.mozilla.geckoview.GeckoResult;
@@ -747,6 +748,7 @@ public class GeckoViewActivity
             mTabSessionManager.setTabObserver(sExtensionManager);
 
             sGeckoRuntime.getWebExtensionController().setDebuggerDelegate(sExtensionManager);
+            sGeckoRuntime.setAutocompleteStorageDelegate(new ExampleAutocompleteStorageDelegate());
 
             // `getSystemService` call requires API level 23
             if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
@@ -1528,6 +1530,8 @@ public class GeckoViewActivity
             Log.i(LOGTAG, "History state updated");
         }
     }
+
+    private class ExampleAutocompleteStorageDelegate implements Autocomplete.StorageDelegate {}
 
     private class ExampleContentDelegate implements GeckoSession.ContentDelegate {
         @Override